# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workflow Service",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://windmill.distribute.org"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WorkflowMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "createdForBrandId": {
            "type": "string",
            "nullable": true
          },
          "featureSlug": {
            "type": "string",
            "description": "Feature slug for grouping and naming."
          },
          "signature": {
            "type": "string"
          },
          "signatureName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "createdForBrandId",
          "featureSlug",
          "signature",
          "signatureName"
        ],
        "description": "Workflow metadata."
      },
      "EmailStats": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "number",
            "description": "Total emails sent."
          },
          "delivered": {
            "type": "number",
            "description": "Total emails delivered."
          },
          "opened": {
            "type": "number",
            "description": "Total emails opened."
          },
          "clicked": {
            "type": "number",
            "description": "Total link clicks."
          },
          "replied": {
            "type": "number",
            "description": "Total replies received."
          },
          "bounced": {
            "type": "number",
            "description": "Total emails bounced."
          },
          "unsubscribed": {
            "type": "number",
            "description": "Total unsubscribes."
          },
          "recipients": {
            "type": "number",
            "description": "Total unique recipients."
          }
        },
        "required": [
          "sent",
          "delivered",
          "opened",
          "clicked",
          "replied",
          "bounced",
          "unsubscribed",
          "recipients"
        ],
        "description": "Aggregated transactional email stats across all runs."
      },
      "WorkflowStats": {
        "type": "object",
        "properties": {
          "totalCostInUsdCents": {
            "type": "number",
            "description": "Total cost across all completed runs of this workflow."
          },
          "totalOutcomes": {
            "type": "number",
            "description": "Total replies or clicks (depending on objective) across all runs."
          },
          "costPerOutcome": {
            "type": "number",
            "nullable": true,
            "description": "Cost per reply or cost per click in USD cents. Null if no outcomes yet."
          },
          "completedRuns": {
            "type": "number",
            "description": "Number of completed runs used in the calculation."
          },
          "email": {
            "type": "object",
            "properties": {
              "transactional": {
                "$ref": "#/components/schemas/EmailStats"
              },
              "broadcast": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/EmailStats"
                  },
                  {
                    "description": "Aggregated broadcast email stats across all runs."
                  }
                ]
              }
            },
            "required": [
              "transactional",
              "broadcast"
            ],
            "description": "Detailed email engagement stats aggregated across the upgrade chain."
          }
        },
        "required": [
          "totalCostInUsdCents",
          "totalOutcomes",
          "costPerOutcome",
          "completedRuns",
          "email"
        ],
        "description": "Aggregated performance stats for this workflow."
      },
...
      "RankedWorkflowObjective": {
        "type": "string",
        "enum": [
          "replies",
          "clicks"
        ],
        "default": "replies",
        "description": "Which metric to optimize for. Defaults to 'replies'."
      },
      "RankedWorkflowGroupBy": {
        "type": "string",
        "enum": [
          "feature",
          "brand"
        ],
        "description": "Group results by section or brand. When omitted, returns a flat ranked list."
      },
      "BestWorkflowRecord": {
        "type": "object",
        "nullable": true,
        "properties": {
          "workflowId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the workflow holding the record."
          },
          "workflowName": {
            "type": "string",
            "description": "Name of the workflow."
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "description": "Stable display name of the workflow family."
          },
          "createdForBrandId": {
            "type": "string",
            "nullable": true,
            "description": "Brand ID that created this workflow (creation context, not execution brand)."
          },
          "value": {
            "type": "number",
            "description": "The record value in USD cents."
          }
        },
        "required": [
          "workflowId",
          "workflowName",
          "displayName",
          "createdForBrandId",
          "value"
        ],
        "description": "Workflow with the lowest cost per email open. Null if no data."
      },
      "BestWorkflowResponse": {
        "type": "object",
        "properties": {
          "bestCostPerOpen": {
            "$ref": "#/components/schemas/BestWorkflowRecord"
          },
          "bestCostPerReply": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BestWorkflowRecord"
              },
              {
                "description": "Workflow with the lowest cost per reply. Null if no data."
              }
            ]
          }
        },
        "required": [
          "bestCostPerOpen",
          "bestCostPerReply"
        ]
      },
      "BestWorkflowBy": {
        "type": "string",
        "enum": [
          "workflow",
          "brand"
        ],
        "default": "workflow",
        "description": "Granularity: best workflow or best brand. Defaults to 'workflow'."
      },
      "PublicRankedWorkflowItem": {
        "type": "object",
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/WorkflowMetadata"
          },
          "stats": {
            "$ref": "#/components/schemas/WorkflowStats"
          }
        },
        "required": [
          "workflow",
          "stats"
        ]
      },
      "PublicRankedWorkflowResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicRankedWorkflowItem"
            },
            "description": "Workflows ranked by performance, best first."
          }
        },
        "required": [
          "results"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/public/workflows/ranked": {
      "get": {
        "summary": "Public: Get workflows ranked by cost-per-outcome",
        "description": "Public version of GET /workflows/ranked. No authentication required. Returns workflows ranked by performance with stats, but without DAG details. Stats are aggregated across the full upgrade chain. Use `groupBy=section` to group by category-channel-audienceType, or `groupBy=brand` to group by brandId. Use `brandId` to filter by a specific brand.",
        "tags": [
          "Public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Organization ID. When omitted, searches across all orgs."
            },
            "required": false,
            "description": "Organization ID. When omitted, searches across all orgs.",
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter workflows by brand ID."
            },
            "required": false,
            "description": "Filter workflows by brand ID.",
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter workflows by feature slug."
            },
            "required": false,
            "description": "Filter workflows by feature slug.",
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/RankedWorkflowObjective"
            },
            "required": false,
            "description": "Which metric to optimize for. Defaults to 'replies'.",
            "name": "objective",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "description": "Max workflows per group (when groupBy is set) or total (when flat). Defaults to 10."
            },
            "required": false,
            "description": "Max workflows per group (when groupBy is set) or total (when flat). Defaults to 10.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/RankedWorkflowGroupBy"
            },
            "required": false,
            "description": "Group results by section or brand. When omitted, returns a flat ranked list.",
            "name": "groupBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked workflows (flat list or grouped by section)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRankedWorkflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "External service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/workflows/best": {
      "get": {
        "summary": "Public: Get hero records — best cost-per-open and best cost-per-reply",
        "description": "Public version of GET /workflows/best. No authentication required. Returns the single best workflow for cost-per-open and cost-per-reply across all active workflows. Stats are aggregated across the full upgrade chain. Use `by=brand` to find the best brand instead of the best workflow. Use `brandId` to filter by a specific brand.",
        "tags": [
          "Public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Organization ID. When omitted, searches across all orgs."
            },
            "required": false,
            "description": "Organization ID. When omitted, searches across all orgs.",
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter workflows by brand ID."
            },
            "required": false,
            "description": "Filter workflows by brand ID.",
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/BestWorkflowBy"
            },
            "required": false,
            "description": "Granularity: best workflow or best brand. Defaults to 'workflow'.",
            "name": "by",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Hero records found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BestWorkflowResponse"
                }
              }
            }
          },
          "502": {
            "description": "External service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}